Move Interval to z1
