+++ geo/z1/interval.go
@@ -1,0 +1,7 @@
+package z1
+
+import (
+	"github.com/jensgreen/dux/geo"
+)
+
+type Interval = geo.Interval[int]
