+++ geo/z2/rect.go
@@ -1,9 +1,12 @@
 package z2
 
 import (
+	"math"
+
 	"github.com/jensgreen/dux/geo"
 	"github.com/jensgreen/dux/geo/r1"
 	"github.com/jensgreen/dux/geo/r2"
+	"github.com/jensgreen/dux/geo/z1"
 )
 
 type Rect = geo.Rect[int]
@@ -12,6 +15,13 @@ func RectAsR2[T geo.Number](rect geo.Rect[T]) r2.Rect {
 	return r2.Rect{
 		X: r1.Interval{Lo: float64(rect.X.Lo), Hi: float64(rect.X.Hi)},
 		Y: r1.Interval{Lo: float64(rect.Y.Lo), Hi: float64(rect.Y.Hi)},
+	}
+}
+
+func snapRoundInterval(interval r1.Interval) z1.Interval {
+	return z1.Interval{
+		Lo: int(math.Floor(interval.Lo)),
+		Hi: int(math.Floor(interval.Hi)),
 	}
 }
