+++ geo/z2/rect_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/jensgreen/dux/geo/r1"
 	"github.com/jensgreen/dux/geo/r2"
+	"github.com/jensgreen/dux/geo/z1"
 )
 
 func TestSnapRoundRect_1(t *testing.T) {
@@ -13,8 +14,8 @@ func TestSnapRoundRect_1(t *testing.T) {
 	got := SnapRoundRect(r2.Rect{X: x, Y: y})
 
 	expected := Rect{
-		X: Interval{0, 0},
-		Y: Interval{1, 2},
+		X: z1.Interval{0, 0},
+		Y: z1.Interval{1, 2},
 	}
 	if !expected.Eq(got) {
 		t.Errorf("Expected %+v, got %+v", expected, got)
@@ -23,8 +24,8 @@ func TestSnapRoundRect_1(t *testing.T) {
 
 func TestRect_ContainsCenter(t *testing.T) {
 	rect := Rect{
-		X: Interval{0, 2},
-		Y: Interval{0, 2},
+		X: z1.Interval{0, 2},
+		Y: z1.Interval{0, 2},
 	}
 	ok := rect.ContainsPoint(1, 1)
 	if !ok {
@@ -34,8 +35,8 @@ func TestRect_ContainsCenter(t *testing.T) {
 
 func TestRect_ContainsLoX(t *testing.T) {
 	rect := Rect{
-		X: Interval{0, 2},
-		Y: Interval{0, 2},
+		X: z1.Interval{0, 2},
+		Y: z1.Interval{0, 2},
 	}
 	ok := rect.ContainsPoint(0, 0) && rect.ContainsPoint(0, 1)
 	if !ok {
@@ -44,8 +45,8 @@ func TestRect_ContainsLoX(t *testing.T) {
 }
 func TestRect_ContainsLoY(t *testing.T) {
 	rect := Rect{
-		X: Interval{0, 2},
-		Y: Interval{0, 2},
+		X: z1.Interval{0, 2},
+		Y: z1.Interval{0, 2},
 	}
 	ok := rect.ContainsPoint(0, 0) && rect.ContainsPoint(1, 0)
 	if !ok {
@@ -55,8 +56,8 @@ func TestRect_ContainsLoY(t *testing.T) {
 
 func TestRect_NotContainsHiX(t *testing.T) {
 	rect := Rect{
-		X: Interval{0, 2},
-		Y: Interval{0, 2},
+		X: z1.Interval{0, 2},
+		Y: z1.Interval{0, 2},
 	}
 	ok := !(rect.ContainsPoint(2, 0) && rect.ContainsPoint(2, 1) && rect.ContainsPoint(2, 2))
 	if !ok {
@@ -65,8 +66,8 @@ func TestRect_NotContainsHiX(t *testing.T) {
 }
 func TestRect_NotContainsHiY(t *testing.T) {
 	rect := Rect{
-		X: Interval{0, 2},
-		Y: Interval{0, 2},
+		X: z1.Interval{0, 2},
+		Y: z1.Interval{0, 2},
 	}
 	ok := !(rect.ContainsPoint(0, 2) && rect.ContainsPoint(1, 2) && rect.ContainsPoint(2, 2))
 	if !ok {
@@ -76,8 +77,8 @@ func TestRect_NotContainsHiY(t *testing.T) {
 
 func TestRect_NotContainsLower(t *testing.T) {
 	rect := Rect{
-		X: Interval{0, 2},
-		Y: Interval{0, 2},
+		X: z1.Interval{0, 2},
+		Y: z1.Interval{0, 2},
 	}
 	ok := !(rect.ContainsPoint(-1, 0) && rect.ContainsPoint(0, -1) && rect.ContainsPoint(-1, -1))
 	if !ok {
@@ -87,8 +88,8 @@ func TestRect_NotContainsLower(t *testing.T) {
 
 func TestRect_NotContainsHigher(t *testing.T) {
 	rect := Rect{
-		X: Interval{0, 2},
-		Y: Interval{0, 2},
+		X: z1.Interval{0, 2},
+		Y: z1.Interval{0, 2},
 	}
 	ok := !(rect.ContainsPoint(3, 0) && rect.ContainsPoint(0, 3) && rect.ContainsPoint(3, 3))
 	if !ok {
